Add files via upload
A travelling game developed during a 24-hour hackathon.

The aim of the game is to complete all three levels successfully by continuously running whilst jumping to avoid obstacles, 
however it is not as easy as it sounds. As the levels progress, so does the speed of the game, ultimately increasing the chances of a collision!

The game will take the user through three cities, each is a level and is in the following order:
- Blackburn (UK)
- Birmingham (UK)
- Milan (Italy)
Each level is the hometown of a developer which took part in the creation of the game, 
the character in each level is a visual representation of the developer. 
All developers now live in Manchester (UK) and thus, that is the end destination for each level.
Only upon completion of all levels will the user be able to see that all the characters have arrived safely in Manchester.